Completed MetaTile environment. Generalisation of fixtures for living entity.

+++ core/src/com/astetyne/expirium/client/gui/roots/game/DeathRoot.java
@@ -19,7 +19,7 @@ public class DeathRoot extends WidgetGroup implements GameRootable {
     private final Label timeRemainingLabel;
     private final boolean firstDeath;
 
-    public DeathRoot(boolean firstDeath, int daysSurvived) {
+    public DeathRoot(boolean firstDeath, long daysSurvived) {
 
         this.firstDeath = firstDeath;
